requirements.txt
sqlalchemy
flask
psutil
python-telegram-bot
dotenv